updated utils and response class

+++ src/main/java/com/abha/sharedlibrary/aums/response/SignupResponse.java
@@ -10,5 +10,5 @@ import lombok.Setter;
 @Getter
 @Setter
 public class SignupResponse {
-  private Integer subscriberId;
+  private Long subscriberId;
 }
